% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convDir.R
\name{convDir}
\alias{convDir}
\title{Convert a directory of raw DICOM images to material classes}
\usage{
convDir(directory, upperLim = 3045, lowerLim = -1024, 
airHU = -850.3233, airSD = 77.6953, 
SiHU = 271.7827, SiSD = 39.2814,
glassHU = 1345.0696, glassSD = 45.4129,
waterHU = 63.912, waterSD = 14.1728,
densities = c(0.0012, 1, 1.23, 2.2))
}
\arguments{
\item{directory}{folder of DICOM images(raw values)}

\item{upperLim}{upper bound cutoff for pixels (Hounsfield Units)}

\item{lowerLim}{lower bound cutoff for pixels (Hounsfield Units)}

\item{airHU}{mean value for air-filled calibration rod (Hounsfield Units)}

\item{airSD}{standard deviation for air-filled calibration rod}

\item{SiHU}{mean value for colloidal silica calibration rod}

\item{SiSD}{standard deviation for colloidal Si calibration rod}

\item{glassHU}{mean value for glass calibration rod}

\item{glassSD}{standard deviation for glass calibration rod}

\item{waterHU}{mean value for water filled calibration rod}

\item{waterSD}{standard deviation for water filled calibration rod}

\item{densities}{numeric vector of known cal rod densities. Format must be c(air, water, Si, glass)}
}
\value{
value \code{convDir} returns a dataframe with one row per CT slice. Values returned are the area and volume of 7 material classes: gas, peat, roots and rhizomes, rock and shell, fine mineral particles, sand, and water.
}
\description{
Calculates the area and volume of material classes for each CT slice in a directory
}
\details{
Calculates the area and volume of material classes for each CT slice in a directory. Unlike \code{\link{conv}}, \code{\link{convDir}} accepts a folder of raw values and makes the conversion to Hounsfield Units using the metadata associated with the DICOM images.
}
\examples{
\dontrun{ 
CTdir <- "C:/RDATA/CT/426/"
materials <- convDir(CTdir)

# plot using "ggplot" package after transforming with "reshape2" package
mass.long <- reshape2::melt(materials, id.vars = c("depth"), 
   measure.vars = grep(".g", names(materials)))
ggplot2::ggplot(data = mass.long, ggplot2::aes(y = -depth, x = value, 
   color = variable)) + ggplot2::geom_point() + ggplot2::theme_classic() + 
   ggplot2::xlab("mass per section (g)") 
}

}
\seealso{
\code{\link{convDir}} is a wrapper for \code{\link{conv}}. \code{\link{rootSizeDir}} operates similarly.
}

