% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_426.R
\docType{data}
\name{core_426}
\alias{core_426}
\title{Ninety-five CT scans from the top of a Spartina alterniflora core}
\format{A list of 95 lists, each with two elements:
\describe{
  \item{hdr}{header information and metadata}
  \item{img}{CT image data}
  ...
}}
\usage{
core_426
}
\description{
CT scans loaded by oro.dicom::readDICOM.
}
\keyword{datasets}

