% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootSize.R
\name{rootSize}
\alias{rootSize}
\title{Root surface area}
\usage{
rootSize(mat.list, pixelA, diameter = c(1, 2, 5, 10, 20),
class.names = diameter,
thickness = 0.625,
airHU = -850.3233,
airSD = 77.6953,
waterHU = 63.912,
waterSD = 14.1728)
}
\arguments{
\item{mat.list}{list of DICOM images for a sediment core}

\item{pixelA}{pixel area (mm2)}

\item{diameter}{an integer vector of diameter cut points. Units are mm (zero is added in automatically).}

\item{class.names}{not used presently}

\item{thickness}{CT image thickness (mm)}

\item{airHU}{mean value for air-filled calibration rod (all rod arguments are in Hounsfield Units)}

\item{airSD}{standard deviation for air-filled calibration rod}

\item{waterHU}{mean value for water-filled calibration rod}

\item{waterSD}{standard deviation for water-filled calibration rod}
}
\value{
value \code{rootSize} returns a dataframe with one row per CT slice. Values returned are the number and surface area of particles in each size class with an upper bound defined in \code{diameter}.
}
\description{
Calculates the number of root/rhizome particles and surface areas, for different size classes
}
\details{
Calculates the number of root/rhizome particles and surface areas, for different size classes
}
\examples{
\dontrun{
data(core_426)
ct.slope <- unique(extractHeader(core_426$hdr, "RescaleSlope"))
ct.int   <- unique(extractHeader(core_426$hdr, "RescaleIntercept")) 
# convert raw units to Hounsfield units
HU_426 <- lapply(core_426$img, function(x) x*ct.slope + ct.int)

rootChars <- rootSize(HU_426, pixelA = 0.244141^2)

# plot using "ggplot" package after transforming with "reshape2" package
area.long <- reshape2::melt(rootChars, id.vars = c("depth"), measure.vars = grep("Area", names(rootChars)))
ggplot2::ggplot(data = area.long, aes(y = -depth, x = value, color = variable)) + 
geom_point() + theme_classic() + xlab("root external surface area per slice (cm2)")
}

}
\seealso{
\code{\link{conv}}
}

