% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootSizeDir.R
\name{rootSizeDir}
\alias{rootSizeDir}
\title{Convert a directory of raw DICOM images to  root particle counts and surface areas}
\usage{
rootSizeDir(directory, diameter = c(1, 2, 5, 10, 20),
class.names = diameter,
airHU = -850.3233,
airSD = 77.6953,
waterHU = 63.912,
waterSD = 14.1728)
}
\arguments{
\item{directory}{directory of DICOM images (raw values)}

\item{diameter}{an integer vector of diameter cut points. Units are mm (zero is added in automatically).}

\item{class.names}{not used presently}

\item{airHU}{mean value for air-filled calibration rod (all rod arguments are in Hounsfield Units)}

\item{airSD}{standard deviation for air-filled calibration rod}

\item{waterHU}{mean value for water-filled calibration rod}

\item{waterSD}{standard deviation for water-filled calibration rod}
}
\value{
value \code{rootSize} returns a dataframe with one row per CT slice. Values returned are the number and surface area of particles in each size class with an upper bound defined in \code{diameter}.
}
\description{
Calculates the number of root/rhizome particles and surface areas, for different size classes
}
\details{
Calculates the number of root/rhizome particles and surface areas, for different size classes. Unlike \code{\link{rootSize}}, \code{\link{rootSizeDir}} accepts a folder of raw values and makes the conversion to Hounsfield Units using the metadata associated with the DICOM images.
}
\examples{
\dontrun{
CTdir <- "C:/RDATA/CT/426/"
rootChars <- rootSizeDir(CTdir)

# plot using "ggplot" package after transforming with "reshape2" package
area.long <- reshape2::melt(rootChars, id.vars = c("depth"), 
   measure.vars = grep("Area", names(rootChars)))
ggplot2::ggplot(data = area.long, ggplot2::aes(y = -depth, x = value, 
   color = variable)) + ggplot2::geom_point() + ggplot2::theme_classic() + 
   ggplot2::xlab("root external surface area per slice (cm2)")
}

}
\seealso{
\code{\link{rootSizeDir}} is a wrapper for \code{\link{rootSize}}. \code{\link{rootSizeDir}} operates similarly.
}

