% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_426.R
\docType{data}
\name{core_426}
\alias{core_426}
\title{Ninety-five CT scans from the top of a Spartina alterniflora core}
\format{A list of 95 matrices, each with two elements: header and image data}
\usage{
data(core_426)
}
\description{
Ninety-five CT scans from the top of a Spartina alterniflora core
}
\keyword{datasets}
