% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_426.R
\docType{data}
\name{core_426}
\alias{core_426}
\title{Three computed tomography scans from a Spartina alterniflora core}
\format{
A list of 3 matrices, each with two elements: header and image data
}
\usage{
data(core_426)
}
\description{
Three computed tomography scans from a Spartina alterniflora core
}
\keyword{datasets}
